DRY up ORM models
- Move models to lib
- Add MIT license to project root
