+++ lib/jukebox_db/README.md
@@ -1,0 +1,3 @@
+# Jukebox DB 
+
+SQLAlchemy ORM models
