+++ lib/jukebox_db/jukebox_db/__init__.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env/python3
+
 import datetime
 import os
 import uuid
@@ -62,3 +64,53 @@ class Follower(Base):
         DateTime, default=datetime.datetime.now, onupdate=datetime.datetime.now
     )
     deleted_at = Column(DateTime)
+
+
+class User(Base):
+    __tablename__ = "users"
+    __table_args__ = {"schema": "_user"}
+
+    id = Column(
+        UUID(as_uuid=True),
+        primary_key=True,
+        default=uuid.uuid4,
+        unique=True,
+        nullable=False,
+    )
+
+    profile = Column(JSON)
+    email = Column(String(250))
+    service_key = Column(String(250))
+    service = Column(String(250))
+
+    created_at = Column(DateTime)
+    updated_at = Column(DateTime)
+    deleted_at = Column(DateTime)
+
+
+class Authorization(Base):
+    __tablename__ = "authorizations"
+    __table_args__ = {"schema": "authorizer"}
+
+    id = Column(
+        UUID(as_uuid=True),
+        primary_key=True,
+        default=uuid.uuid4,
+        unique=True,
+        nullable=False,
+    )
+
+    access_token_expiration = Column(DateTime)
+    access_token = Column(String(250))
+    refresh_token = Column(String(250))
+    scope = Column(String(250))
+    service = Column(String(250))
+
+    created_at = Column(DateTime, default=datetime.datetime.now)
+    updated_at = Column(
+        DateTime, default=datetime.datetime.now, onupdate=datetime.datetime.now
+    )
+    deleted_at = Column(DateTime)
+
+
+__version__ = "0.0.1"
