+++ services/authorizer/workers/make_authorized_request/requirements.txt
@@ -14,6 +14,7 @@ certifi==2020.4.5.1
 chardet==3.0.4
 click==7.1.2
 git+https://github.com/buckmaxwell/jukebox.git@master#egg=async_messenger&subdirectory=lib/async_messenger
+git+https://github.com/buckmaxwell/jukebox.git@record-plays#egg=jukebox_db&subdirectory=lib/jukebox_db
 idna==2.9
 isort==4.3.21
 itsdangerous==1.1.0
