+++ services/authorizer/workers/refresh_authorization/requirements.txt
@@ -3,17 +3,19 @@ arrow==0.15.5
 astroid==2.4.0
 attrs==19.3.0
 black==19.10b0
+Flask-HTTPAuth==4.0.0
+Flask==1.1.2
+Jinja2==2.11.2
+MarkupSafe==1.1.1
+SQLAlchemy==1.3.16
 certifi==2020.4.5.1
 chardet==3.0.4
 click==7.1.2
-Flask==1.1.2
-Flask-HTTPAuth==4.0.0
+git+https://github.com/buckmaxwell/jukebox.git@record-plays#egg=jukebox_db&subdirectory=lib/jukebox_db
 idna==2.9
 isort==4.3.21
 itsdangerous==1.1.0
-Jinja2==2.11.2
 lazy-object-proxy==1.4.3
-MarkupSafe==1.1.1
 mccabe==0.6.1
 pathspec==0.8.0
 pika==1.1.0
@@ -25,7 +27,6 @@ regex==2020.4.4
 requests==2.23.0
 sentry-sdk==0.14.4
 six==1.14.0
-SQLAlchemy==1.3.16
 toml==0.10.0
 typed-ast==1.4.1
 urllib3==1.25.9
