+++ services/player/workers/create_play/requirements.txt
@@ -15,6 +15,7 @@ certifi==2020.4.5.1
 chardet==3.0.4
 click==7.1.2
 decorator==4.4.2
+git+https://github.com/buckmaxwell/jukebox.git@record-plays#egg=jukebox_db&subdirectory=lib/jukebox_db
 idna==2.9
 ipdb==0.13.2
 ipython-genutils==0.2.0
