+++ services/room_settings/requirements.txt
@@ -14,6 +14,7 @@ chardet==3.0.4
 click==7.1.2
 git+https://github.com/buckmaxwell/jukebox.git@master#egg=async_messenger&subdirectory=lib/async_messenger
 git+https://github.com/buckmaxwell/jukebox.git@master#egg=redis_wait&subdirectory=lib/redis_wait
+git+https://github.com/buckmaxwell/jukebox.git@record-plays#egg=jukebox_db&subdirectory=lib/jukebox_db
 gunicorn==20.0.4
 idna==2.9
 isort==4.3.21
