+++ services/typeahead/requirements.txt
@@ -11,6 +11,8 @@ certifi==2020.4.5.1
 chardet==3.0.4
 click==7.1.2
 git+https://github.com/buckmaxwell/jukebox.git@master#egg=spotify_request&subdirectory=lib/spotify_request
+git+https://github.com/buckmaxwell/jukebox.git@record-plays#egg=jukebox_db&subdirectory=lib/jukebox_db
+git+https://github.com/buckmaxwell/jukebox.git@record-plays#egg=jukebox_db&subdirectory=lib/jukebox_db
 gunicorn==20.0.4
 idna==2.9
 isort==4.3.21
